Collectable Tracking and Game State Management
Did some work on game state management by adding public enum to GameManager script, I'll have to play around with it later. Also added global tracking for collectables in game manager. Added level manager script which so far takes care of updating collectables on level entry if they've already been gotten and updating the global collectables on level completion. Edited loadTrigger script to include debug checkbox specifically for me and trying to get collectable tracking down. Edited collectable script to make updating the ui a seperate method and adding a boolean 'collected' which is set to true if the sock has been collected in a previous playthrough of the level