Lots of work with ui manager, game level manager, collectables, timer, level loading etc
Created UIManager to manage the ui elements of the game, added lots of tool tips and stuff so it should be easy to understand. I edited the new player prefab to include the player hit script. I edited the collectables to instead of editing the ui directly, they call the uimanager to perform a function to do so. Added in the timer to the UI and got it counting down in level manager.  Got rudamentary score calulation done in level manager. created a timer script that should work universily. Finally there's a weird bug with level loading where when you reload into the level the camera doesn't follow the player character and also it doesn't seem like time ticks down.