Title Screen and Cutscene
I've added a title screen to the project with buttons to start game, enter an options menu, enter a credits menu, or play in debug mode (all the last one does is set the game progress to being at the end of game and spawns you in the hub with access to every level). I also added a cutscene manager and an example of how cutscenes would work in a scene titles Cutscene. When you press play on the title you get sent to the cutscene and when that's over you get sent to the hub. I also added the garden scene to the build settings