Fixed Grind Rails
I've Fixed grind Rails by making an extra skateboard object that follows around the player, it ignores raycast and the grind rail checks to see if skateboard collided with it. I added this to each level already so it should work. I also made some edits to the street level to make the environmental stuff match the final mesh. Finally added our actual art asset to the star pickup particle. Added lock rotation to ramps, the only issue is that it lets players fly really far and could mess up level design. Right now we're keeping it as it was